requirements/base.txt
analytics-python==1.1.0
Django==1.8.14
django-appconf==0.6
django-compressor==2.0
django_extensions==1.5.5
django-fernet-fields==0.4
django-filter==0.11.0
django-libsass==0.5
django-oscar==1.1.1
django-rest-swagger==0.3.5
django_simple_history==1.6.3
django-threadlocals==0.8
django-waffle==0.11.1
djangorestframework==3.2.3
djangorestframework-jwt==1.8.0
drf-extensions==0.2.8
edx-auth-backends==0.5.2
edx-django-release-util==0.1.0
edx-django-sites-extensions==1.0.0
edx-drf-extensions==1.0.0
edx-ecommerce-worker==0.5.0
edx-opaque-keys==0.3.1
edx-rest-api-client==1.6.0
jsonfield==1.0.3
libsass==0.9.2
ndg-httpsclient==0.4.0
path.py==7.2
paypalrestsdk==1.11.5
premailer==2.9.2
pycountry==1.18
python-dateutil==2.4.2
pytz==2015.7
requests==2.9.1
sailthru-client==2.2.3
suds==0.4
